# docs/openapi.yaml
openapi: 3.1.0
info:
  title: klaxond API
  version: 0.14.26
  summary: Notification bridge backend API
  description: |
    klaxond receives monitoring webhooks, renders notification payloads, manages
    routing/inhibition policy, and exposes a zero-build admin UI.

    Browser sessions use the `klaxond_session` cookie plus `X-Klaxond-CSRF` on
    same-origin mutations. Scoped machine clients should use Bearer tokens.
    Sensitive browser mutations can also return `428 Precondition Required`
    with `X-Klaxond-Reauth: required`; call `/api/auth/reauth` or complete passkey
    reauthentication before retrying.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
paths:
...
  /api/auth/passkey/login/options:
    post:
      tags: [Auth]
      operationId: passkeyLoginOptions
      summary: Create passkey authentication options.
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
                  description: User subject, username or email for standalone passkey login.
                step_up:
                  type: string
                  description: One-time MFA / step-up token issued after primary authentication.
                return_to:
                  type: string
      responses:
        "200":
          description: WebAuthn authentication challenge.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/passkey/login/verify:
    post:
      tags: [Auth]
      operationId: passkeyLoginVerify
      summary: Verify passkey authentication and issue a session cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Passkey session established.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/step-up:
    get:
      tags: [Auth]
      operationId: stepUpPage
      summary: Render the public MFA / step-up page.
      description: The page is reachable before the app session is issued and is guarded by the one-time step-up token.
      security: []
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
        - in: query
          name: return_to
          schema:
            type: string
            default: /status
      responses:
        "200":
          description: MFA / step-up HTML.
          content:
            text/html:
              schema:
                type: string
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/step-up/status:
    get:
      tags: [Auth]
      operationId: getStepUpStatus
      summary: Return the pending MFA / step-up challenge metadata.
      security: []
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pending challenge metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StepUpStatus"
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/step-up/passkey/register/options:
    post:
      tags: [Auth]
      operationId: stepUpPasskeyRegisterOptions
      summary: Create passkey registration options for an MFA / step-up challenge.
      description: Allows first-passkey enrollment only for the user bound to the pending step-up token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepUpPasskeyRegisterRequest"
      responses:
        "200":
          description: WebAuthn registration challenge.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/step-up/passkey/register/verify:
    post:
      tags: [Auth]
      operationId: stepUpPasskeyRegisterVerify
      summary: Verify a step-up passkey registration and issue the app session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Registered passkey plus issued session user.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/step-up/totp/setup/start:
    post:
      tags: [Auth]
      operationId: stepUpTotpSetupStart
      summary: Start TOTP enrollment for an MFA / step-up challenge.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepUpTokenRequest"
      responses:
        "200":
          description: TOTP setup payload scoped to the step-up user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StepUpTotpSetup"
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/step-up/totp/setup/confirm:
    post:
      tags: [Auth]
      operationId: stepUpTotpSetupConfirm
      summary: Confirm TOTP enrollment and issue the app session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepUpTotpConfirmRequest"
      responses:
        "200":
          description: Issued session user and return URL.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/step-up/totp/verify:
    post:
      tags: [Auth]
      operationId: stepUpTotpVerify
      summary: Verify an existing TOTP factor and issue the app session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepUpTotpVerifyRequest"
      responses:
        "200":
          description: Issued session user and return URL.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/status:
    get:
      tags: [Operations]
      operationId: getStatus
      summary: Return runtime status, channel health and recent activity.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Status payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusPayload"
        "4XX":
          $ref: "#/components/responses/ClientError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: klaxond scoped token
      description: API keys and PATs created from `/api/auth/tokens`.
    sessionCookie:
      type: apiKey
      in: cookie
      name: klaxond_session
      description: Signed browser session cookie.
    csrfHeader:
      type: apiKey
      in: header
      name: X-Klaxond-CSRF
      description: Required on browser-originated mutations when using a session cookie.
    basicAuth:
      type: http
      scheme: basic
      description: Username/password auth for non-browser clients in basic or LDAP mode.
...
  responses:
...
    ClientError:
      description: Generic client-side error, including invalid input, missing credentials or authorization failure depending on the active auth mode.
      content:
        text/plain:
          schema:
            type: string
...
  schemas:
    Severity:
      type: string
      enum: [info, warning, critical]
...
    LoginResponse:
      type: object
      required: [ok, return_to, csrf]
      properties:
        ok:
          type: boolean
        return_to:
          type: string
        csrf:
          type: string
...
    StepUpStatus:
      type: object
      required: [ok, factor, reason, return_to, user, passkey_registered, totp_registered]
      properties:
        ok:
          type: boolean
        factor:
          type: string
          enum: [passkey, hardware_key, totp]
        reason:
          type: string
        return_to:
          type: string
        user:
          type: object
          required: [sub, name, email]
          properties:
            sub:
              type: string
            name:
              type: string
            email:
              type: string
        passkey_registered:
          type: boolean
        totp_registered:
          type: boolean
    StepUpTokenRequest:
      type: object
      required: [step_up]
      properties:
        step_up:
          type: string
    StepUpPasskeyRegisterRequest:
      allOf:
        - $ref: "#/components/schemas/StepUpTokenRequest"
        - type: object
          properties:
            name:
              type: string
    StepUpTotpSetup:
      type: object
      required: [ok, request_id, secret, otpauth_uri]
      properties:
        ok:
          type: boolean
        request_id:
          type: string
        secret:
          type: string
        otpauth_uri:
          type: string
          format: uri
    StepUpTotpConfirmRequest:
      type: object
      required: [request_id, code]
      properties:
        request_id:
          type: string
        code:
          type: string
    StepUpTotpVerifyRequest:
      allOf:
        - $ref: "#/components/schemas/StepUpTokenRequest"
        - type: object
          required: [code]
          properties:
            code:
              type: string
...
    StatusPayload:
      type: object
      additionalProperties: true
      properties:
        version:
          type: string
        cascade_enabled:
          type: boolean
        channels:
          type: object
        recent_deliveries:
          type: array
          items:
            $ref: "#/components/schemas/Delivery"
    Delivery:
      type: object
      additionalProperties: true
      properties:
        ts:
          type: number
          format: double
          description: Unix timestamp in seconds.
        source:
          type: string
        severity:
          $ref: "#/components/schemas/Severity"
        title:
          type: string
        channel:
          type: string
        suppressed_by:
          type: string
        ok:
          type: boolean
...
security:
  - bearerAuth: []
  - sessionCookie: []